+++ src/main/java/ch/ethz/idsc/amodeus/routing/DefaultAStarLMRouter.java
@@ -54,7 +54,7 @@ public class DefaultAStarLMRouter implements AVRouter {
         @Override
         public AVRouter createRouter(RouterConfig routerConfig, Network network) {
             return new DefaultAStarLMRouter(DefaultParallelLeastCostPathCalculator.//
-                    create((int) config.getNumberOfThreads(), new FastAStarLandmarksFactory(), network, //
+                    create(config.getNumberOfThreads(), new FastAStarLandmarksFactory(), network, //
                             new OnlyTimeDependentTravelDisutilityFixed(travelTime), travelTime));
 
         }
