+++ src/test/java/ch/ethz/idsc/amodeus/testutils/TestPreparer.java
@@ -49,7 +49,7 @@ public class TestPreparer {
         Scenario scenario = ScenarioUtils.loadScenario(config);
         GeneratorConfig avGeneratorConfig = //
                 avConfigGroup.getOperatorConfigs().values().iterator().next().getGeneratorConfig();
-        int numRt = (int) avGeneratorConfig.getNumberOfVehicles();
+        int numRt = avGeneratorConfig.getNumberOfVehicles();
         int endTime = (int) config.qsim().getEndTime();
 
         // 1) cut network (and reduce population to new network)
