+++ src/test/java/ch/ethz/idsc/amodeus/traveldata/PopulationToolsTestVN3.java
@@ -61,7 +61,7 @@ public class PopulationToolsTestVN3 {
         AVConfigGroup avCg = new AVConfigGroup();
         Config config = ConfigUtils.loadConfig(configFile.getAbsolutePath(), avCg);
         GeneratorConfig genConfig = avCg.getOperatorConfigs().values().iterator().next().getGeneratorConfig();
-        int numRt = (int) genConfig.getNumberOfVehicles();
+        int numRt = genConfig.getNumberOfVehicles();
         endTime = (int) config.qsim().getEndTime();
         Scenario scenario = ScenarioUtils.loadScenario(config);
         network = scenario.getNetwork();
