+++ src/test/java/ch/ethz/idsc/amodeus/virtualnetwork/KMeansVirtualNetworkCreatorTest.java
@@ -43,7 +43,7 @@ public class KMeansVirtualNetworkCreatorTest {
         AVConfigGroup avCg = new AVConfigGroup();
         Config config = ConfigUtils.loadConfig(configFile.getAbsolutePath(), avCg);
         GeneratorConfig genConfig = avCg.getOperatorConfigs().values().iterator().next().getGeneratorConfig();
-        int numRt = (int) genConfig.getNumberOfVehicles();
+        int numRt = genConfig.getNumberOfVehicles();
         int endTime = (int) config.qsim().getEndTime();
         Scenario scenario = ScenarioUtils.loadScenario(config);
         Network network = scenario.getNetwork();
